Add IS_SIGNED_TYPE constant
Previously, we checked Self::MIN == Self::ZERO in the construct_bigint
macro, when we needed different behaviour for signed and unsigned
types. Replace those with IS_SIGNED_TYPE constant, for clarity.

This also seems to affect the generated code slightly. Looking at the
code generated for i1024::wrapping_add, for example, it's slighty
smaller with this. ISTM that the compiler does not always optimize
away the Self::MIN == Self::ZERO as you might think. I'm also seeing a
17% improvement in the u1024_mul benchmark, although I'm not certain
if that's just an artifact of code moving around or a real effect.