Rewrite signed cmp() in a more performant way
For each word, use the u64::cmp() or i64::cmp() function. This leads
to generated code that, for each pair of words, first checks whether
the words are equal, and only if they're not, checks which one is
larger. The generated code is shorter, which is generally good, and
performs fewer comparisons when the inputs are equal, or if the have
many leading words that are equal. The only case where this
theoretically loses is if self < other, and they differ at the first
word, and even in that case the difference is not big, two comparisons
instead of one. I was not able to measure a difference in that case
either.